refactor: extract instruction bridge flows
